LayoutTests/imported/w3c/web-platform-tests/html/browsers/browsing-the-web/history-traversal/browsing_context_name_cross_origin_2-expected.txt

Blocked access to external URL http://www.localhost:8800/html/browsers/browsing-the-web/history-traversal/browsing_context_name-2.html
Blocked access to external URL http://www.localhost:8800/html/browsers/browsing-the-web/history-traversal/browsing_context_name-2.html

FAIL Restoring window.name on cross-origin history traversal assert_equals: After navigation 1 expected "test3" but got "test"

Step 1 http://localhost:8800/html/browsers/browsing-the-web/history-traversal/browsing_context_name-1.html
Step 2 http://localhost:8800/html/browsers/browsing-the-web/history-traversal/browsing_context_name-1.html
Step 3 http://localhost:8800/html/browsers/browsing-the-web/history-traversal/browsing_context_name-1.html
Step 4 http://localhost:8800/html/browsers/browsing-the-web/history-traversal/browsing_context_name-1.html
